refactor User to repository
