+++ app/Services/UserService.php
@@ -16,5 +16,13 @@ class UserService
         $this->repository = $repository;
     }
 
-    
+    public function create($data)
+    {
+        $this->repository->create($data);
+    }
+
+    public function findWhere($where, bool $getFirst = false)
+    {
+        $this->repository->create($data);
+    }
 }
